Move FlyStr and FlyByteStr into submodules
